Refactored Row component.
