+++ src/App.js
@@ -3,7 +3,7 @@ import React from "react";
 import { useSelector, useDispatch } from 'react-redux';
 import styled from "styled-components";
 
-import Row from './components/row.component';
+import Row from './components/row/row.component';
 import {
   selectBoxesInputted,
   selectErrors,
