+++ src/components/row/row.styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+import { BOARD_ORDER } from "../../constants";
+
+export const RowContainer = styled.div`
+display: grid;
+grid-template-columns: repeat(${BOARD_ORDER}, 180px);
+`;
